Sped up SVD-dependent regression.
Regression scripts were slow in part because of the need to process each data vector separately, each time recomputing the SVD of the features. The updated scripts allow multiple vectors to be input to regression scripts and the SVD thus only needs to be computed once for the whole batch.